+++ app/views/sheets/show.blade.php
@@ -17,15 +17,17 @@
                 </ul>
                 <div class="row">
                     <div class="col-xs-12 sheetcolbtns">
-                        <a class="btn btn-edit" role="button" href="{{{ action('SheetsController@edit', $sheet->id) }}}">Edit</a>
-                        {{ Form::model($sheet, array('action' => array('SheetsController@destroy', $sheet->id), 'method' => 'DELETE', 'class' => 'deleteform')) }}
-                            <button class="btn btn-danger deletebtn" type="submit">Delete</button>
-                        {{ Form::close() }}
+                        @if(Auth::user()->id == $sheet->user_id)
+                            <a class="btn btn-edit" role="button" href="{{{ action('SheetsController@edit', $sheet->id) }}}">Edit</a>
+                            {{ Form::model($sheet, array('action' => array('SheetsController@destroy', $sheet->id), 'method' => 'DELETE', 'class' => 'deleteform')) }}
+                                <button class="btn btn-danger deletebtn" type="submit">Delete</button>
+                            {{ Form::close() }}
+                        @endif
                     </div>
                     <div class="col-xs-12 sheetcolbtns">
                         <a class="btn btn-back" role="button" href="{{{ action('SheetsController@index') }}}">Back</a>
                         <span class="btn btn-standard" role="button" id="toggle-btn">Toggle</span>
-                        <a class="btn btn-success" href="{{{action('CollaborationController@showinvitesheet', array($sheet->id))}}}">Invite</a>
+                    
 
                         <h4>Votes:
                             <span id="voteUpCounts"> {{ $sheet->voteUpCount() }}</span> | <span id="voteDownCounts">-{{ $sheet->voteDownCount() }}</span>
@@ -35,12 +37,6 @@
                         
                         <button  id="voteDown" class="btn btn-standard" data-sheet-id="{{ $sheet->id }}" data-vote="0"><span class="glyphicon glyphicon-triangle-bottom arrowBig" aria-hidden="true"></span></button>
 
-                        <h4>People collaborating on this sheet:</h4>
-                        <ul>
-                            @foreach($collaborators as $guest)
-                            <li>{{{$guest}}}</li>
-                            @endforeach
-                        </ul>
                     </div>
 
                 </div>
@@ -68,7 +64,7 @@
                 @foreach($comments as $comment)
                 <h5>@if($comment['commenter'] != null){{{$comment['commenter']}}} says: @else {{{$sheet->user->firstname . ' ' . $sheet->user->lastname}}} says: @endif</h5>
                 <textarea class="commentarea" type="disabled" disabled rows="7" cols="50">{{{$comment['comment']}}}</textarea>
-                @if((Auth::user()->firstname . Auth::user()->lastname) == $comment['commenter'] || Auth::user() == $sheet->user)
+                @if((Auth::user()->firstname . ' ' . Auth::user()->lastname) == $comment['commenter'] || Auth::user() == $sheet->user)
                     <a class="btn btn-edit" href="{{{action('CollaborationController@showeditcommentsheet', array($sheet->id, $comment['id']))}}}">Edit</a>
                 @endif
                 @endforeach
